microservices/dash-service/requirements.txt
nameko
nameko-redis
kafka-python